+++ project/myPT4/products.php
@@ -20,7 +20,7 @@ include_once 'products_crud.php';
       <script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
       <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
     <![endif]-->
-      <link rel="stylesheet" type="text/css" href="body.css">
+      <link rel="stylesheet" type="text/css" href="css/body.css">
       <style>
       input[type="file"] {
         display: none;
@@ -36,40 +36,43 @@ include_once 'products_crud.php';
     <a href="orders.php">Orders</a>
     <hr> -->
     <?php include_once 'nav_bar.php'; ?>
-    <div class="container-fluid">
 
-      <div class="row">
-        <div class="col-xs-12 col-sm-8 col-sm-offset-2 col-md-6 col-md-offset-3">
-          <?php if (isset($_SESSION['error'])) {
-            echo "<div class='alert alert-danger' role='alert'>{$_SESSION['error']}</div>";
-            unset($_SESSION['error']);
-          }  elseif(isset($_SESSION['success'])){
-            echo "<div class='alert alert-success' role='alert'>{$_SESSION['success']}</div>";
-            unset($_SESSION['success']);
-          }
-          ?>
-          
+    <?php
+    if (isset($_SESSION['user']) && $_SESSION['user']['FLD_ROLE'] == 'admin') {
+      ?>
+      <div class="container-fluid">
+        <div class="row">
+          <div class="col-xs-12 col-sm-8 col-sm-offset-2 col-md-6 col-md-offset-3">
+            <?php if (isset($_SESSION['error'])) {
+              echo "<div class='alert alert-danger' role='alert'>{$_SESSION['error']}</div>";
+              unset($_SESSION['error']);
+            }  elseif(isset($_SESSION['success'])){
+              echo "<div class='alert alert-success' role='alert'>{$_SESSION['success']}</div>";
+              unset($_SESSION['success']);
+            }
+            ?>
+            
 
-          <div class="page-header">
+            <div class="page-header">
 
-            <h2>Add New Product</h2>
-          </div>
-          <div class="row">
-            <form action="<?php echo($_SERVER['REQUEST_URI']); ?>" method="post" class="form-horizontal" enctype="multipart/form-data">
-              <!-- Image -->
-              <div class="col-md-4">
-                <div class="thumbnail dark-1" style="background-color: #696969;">
-                  <img src="products/<?php echo (isset($_GET['edit']) ? $editrow['FLD_IMAGE'] : '') ?>" onerror="this.onerror=null;this.src='products/noimage.png';" id="pphoto" alt="Product Image" style="width: 100%;height: 100%;">
-                  <div class="caption text-center">
-                    <h3 id="productImageTitle" style="word-break: break-all;color:#fbc306;"><?php echo (isset($_GET['edit']) && $editrow['FLD_IMAGE'] !== '' ? explode('.', $editrow['FLD_IMAGE'])[0] : 'Product Image') ?></h3> <!-- explode like split, it will remove the element after . -->
-                    <p>
-                      <label class="btn btn-primary">
-                        <input type="file" accept="images/*" name="fileToUpload" id="inputImage" onchange="loadFile(event);" />
-                        <input type="hidden" name="filename" value="<?php echo $editrow['FLD_IMAGE']; ?>">
-                        <i class="fa fa-cloud-upload"></i>Upload
+              <h2>Add New Product</h2>
+            </div>
+            <div class="row">
+              <form action="<?php echo($_SERVER['REQUEST_URI']); ?>" method="post" class="form-horizontal" enctype="multipart/form-data">
+                <!-- Image -->
+                <div class="col-md-4">
+                  <div class="thumbnail dark-1" style="background-color: #696969;">
+                    <img src="products/<?php echo (isset($_GET['edit']) ? $editrow['FLD_IMAGE'] : '') ?>" onerror="this.onerror=null;this.src='products/noimage.png';" id="pphoto" alt="Product Image" style="width: 100%;height: 100%;">
+                    <div class="caption text-center">
+                      <h3 id="productImageTitle" style="word-break: break-all;color:#fbc306;"><?php echo (isset($_GET['edit']) && $editrow['FLD_IMAGE'] !== '' ? explode('.', $editrow['FLD_IMAGE'])[0] : 'Product Image') ?></h3> <!-- explode like split, it will remove the element after . -->
+                      <p>
+                        <label class="btn btn-primary">
+                          <input type="file" accept="images/*" name="fileToUpload" id="inputImage" onchange="loadFile(event);" />
+                          <input type="hidden" name="filename" value="<?php echo $editrow['FLD_IMAGE']; ?>">
+                          <i class="fa fa-cloud-upload"></i>Upload
 
-                      </label>
-                      <?php
+                        </label>
+                        <?php
                     /*if (isset($_GET['edit']) && $editrow['FLD_IMAGE'] != '') {
                       echo '<a href="#" class="btn btn-danger disabled" role="button">Delete</a>';
                     }*/
@@ -146,7 +149,8 @@ include_once 'products_crud.php';
       </div>
     </div>
 
-
+  <?php } ?>
+  <div class="container">
     <div class="row">
       <div class="col-xs-12 col-sm-10 col-sm-offset-1 col-md-8 col-md-offset-2">
         <div class="page-header">
@@ -193,9 +197,13 @@ include_once 'products_crud.php';
               <td><?php echo $readrow['FLD_COLOUR']; ?></td>
               <td>
                 <a href="products_details.php?pid=<?php echo $readrow['FLD_PRODUCT_ID']; ?>" class="btn btn-warning btn-xs" role="button">Details</a>
-                <a href="products.php?edit=<?php echo $readrow['FLD_PRODUCT_ID']; echo (isset($_GET['page']) ? '&page='.$_GET['page'] : ''); ?>" class="btn btn-success btn-xs" role="button"> Edit </a>
-                <!-- <a href="products.php?edit=<?php echo $readrow['FLD_PRODUCT_ID']; ?>" class="btn btn-success btn-xs" role="button">Edit</a> -->
-                <a href="products.php?delete=<?php echo $readrow['FLD_PRODUCT_ID']; ?>" onclick="return confirm('Are you sure to delete?');" class="btn btn-danger btn-xs" role="button">Delete</a>
+                <?php
+                if (isset($_SESSION['user']) && $_SESSION['user']['FLD_ROLE'] == 'admin') {
+                  ?>
+                  <a href="products.php?edit=<?php echo $readrow['FLD_PRODUCT_ID']; echo (isset($_GET['page']) ? '&page='.$_GET['page'] : ''); ?>" class="btn btn-success btn-xs" role="button"> Edit </a>
+                  <!-- <a href="products.php?edit=<?php echo $readrow['FLD_PRODUCT_ID']; ?>" class="btn btn-success btn-xs" role="button">Edit</a> -->
+                  <a href="products.php?delete=<?php echo $readrow['FLD_PRODUCT_ID']; ?>" onclick="return confirm('Are you sure to delete?');" class="btn btn-danger btn-xs" role="button">Delete</a>
+                <?php } ?>
               </td>
             </tr>
             <?php
@@ -218,47 +226,49 @@ include_once 'products_crud.php';
     </table>
   </div>
 </div>
+</div>
 
-<div class="row">
-  <div class="col-xs-12 col-sm-10 col-sm-offset-1 col-md-8 col-md-offset-2">
-    <nav>
-      <ul class="pagination">
-        <?php
-        try {
-          $conn = new PDO("mysql:host=$servername;dbname=$dbname", $username, $password);
-          $conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-          $stmt = $conn->prepare("SELECT * FROM tbl_products_a174088_pt2");
-          $stmt->execute();
-          $result = $stmt->fetchAll();
-          $total_records = count($result);
-        }
-        catch(PDOException $e){
-          echo "Error: " . $e->getMessage();
-        }
-        $total_pages = ceil($total_records / $per_page);
-        ?>
-        <?php if ($page==1) { ?>
-          <li class="disabled"><span aria-hidden="true">«</span></li>
-        <?php } else { ?>
-          <li><a href="products.php?page=<?php echo $page-1 ?>" aria-label="Previous"><span aria-hidden="true">«</span></a></li>
+<div class="container">
+  <div class="row">
+    <div class="col-xs-12 col-sm-10 col-sm-offset-1 col-md-8 col-md-offset-2">
+      <nav>
+        <ul class="pagination">
           <?php
-        }
-        for ($i=1; $i<=$total_pages; $i++)
-          if ($i == $page)
-            echo "<li class=\"active\"><a href=\"products.php?page=$i\">$i</a></li>";
-          else
-            echo "<li><a href=\"products.php?page=$i\">$i</a></li>";
+          try {
+            $conn = new PDO("mysql:host=$servername;dbname=$dbname", $username, $password);
+            $conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+            $stmt = $conn->prepare("SELECT * FROM tbl_products_a174088_pt2");
+            $stmt->execute();
+            $result = $stmt->fetchAll();
+            $total_records = count($result);
+          }
+          catch(PDOException $e){
+            echo "Error: " . $e->getMessage();
+          }
+          $total_pages = ceil($total_records / $per_page);
           ?>
-          <?php if ($page==$total_pages) { ?>
-            <li class="disabled"><span aria-hidden="true">»</span></li>
+          <?php if ($page==1) { ?>
+            <li class="disabled"><span aria-hidden="true">«</span></li>
           <?php } else { ?>
-            <li><a href="products.php?page=<?php echo $page+1 ?>" aria-label="Previous"><span aria-hidden="true">»</span></a></li>
-          <?php } ?>
-        </ul>
-      </nav>
+            <li><a href="products.php?page=<?php echo $page-1 ?>" aria-label="Previous"><span aria-hidden="true">«</span></a></li>
+            <?php
+          }
+          for ($i=1; $i<=$total_pages; $i++)
+            if ($i == $page)
+              echo "<li class=\"active\"><a href=\"products.php?page=$i\">$i</a></li>";
+            else
+              echo "<li><a href=\"products.php?page=$i\">$i</a></li>";
+            ?>
+            <?php if ($page==$total_pages) { ?>
+              <li class="disabled"><span aria-hidden="true">»</span></li>
+            <?php } else { ?>
+              <li><a href="products.php?page=<?php echo $page+1 ?>" aria-label="Previous"><span aria-hidden="true">»</span></a></li>
+            <?php } ?>
+          </ul>
+        </nav>
+      </div>
     </div>
   </div>
-
 
   <!-- </center> -->
 </div>
